+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Versions
 
+## x.x.x
+    * Fix safe area blocked if y offset used. (https://github.com/AppLovin/AppLovin-MAX-React-Native/issues/59)
 ## 3.0.0
     * Lock to Android SDK 11.3.0 and iOS SDK 11.3.0.
 ## 2.5.3